fixed some merg issues
